+++ src/server/game/Chat/Chat.cpp
@@ -673,7 +673,7 @@ void ChatHandler::FillMessageData(WorldPacket* data, WorldSession* session, uint
             }
             *data << uint32(messageLength);
             *data << message;
-            *data << uint8(0);
+            *data << uint16(0);
 
             if (type == CHAT_MSG_RAID_BOSS_WHISPER || type == CHAT_MSG_RAID_BOSS_EMOTE)
             {
